fix(datepicker): add translations from momentJS dependency
What's the issue?
Ant Desing uses Moment js for Dates, so i18n in Moment JS, for some contents of the picker shown (Month, days of week, etc...) needs to be specified even if ConfigProvider from Antd knows the language used.
How was resolved?
Create datePickerUtils.js to extract all the logic related to i18n of DatePicker and include translations from MomentJS to translate some contents of the picker